symbol: Allow different path names when loading symbol files
Symbol files contain pathname of the ELF image and its build-id.  The
file name is a basename of the image but the pathname can be little
bit different.  If the build-id matches, we can allow different
pathname.  This can be useful when stripped binaries are saved in a
different location.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>